requirements.txt
openai-whisper
pyaudio
google-generativeai
gtts
pydub
pandas
openpyxl
sounddevice
